переносит функцию генерации числа в util.js
